backend/requirements.txt
# Django core
Django>=5.2,<6.0

# REST API
djangorestframework>=3.15,<4.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Static files (production)
whitenoise>=6.6,<7.0

# Environment management
python-dotenv>=1.0,<2.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
jsonschema>=4.21,<5.0

# Production server
gunicorn>=21.0,<23.0
